backend/requirements.txt
fastapi
uvicorn
python-multipart
openai
anthropic
pypdf
python-docx
python-dotenv
boto3
supabase
numpy
striprtf
textract
